Correções

+++ src/main/java/br/com/postechfiap/fiap_produto_service/dto/ProdutoResponse.java
@@ -3,10 +3,12 @@ package br.com.postechfiap.fiap_produto_service.dto;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
+import java.math.BigDecimal;
+
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record ProdutoResponse(
         Long id,
         String nome,
         String sku,
-        Double preco
+        BigDecimal preco
 ) {}
